requirements.txt
asgiref==3.2.10
certifi==2022.5.18
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-cors-headers==3.11.0
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
numpy==1.23.2
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.3
protobuf==3.20.1
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.1